Implement shared Card with BfR-Logo
    - for HomePage and ErrorPage

Ticket: 203
