+++ src/app/Main/Body/ErrorPage/Container-PageNotFound.component.tsx
@@ -1,33 +1,15 @@
-/** @jsx jsx */
-import { css, jsx } from "@emotion/core";
-import { Card, CardMedia, makeStyles } from "@material-ui/core";
-import { ErrorPageTextContentComponent } from "./PageNotFound-TextContent.component";
-
-const logoStyle = css`
-    height: auto;
-    width: auto;
-`;
-const useStyles = makeStyles({
-    root: {
-        padding: "1em",
-        margin: "2.5em auto",
-        width: "50%",
-    },
-});
+import React from "react";
+import { useTranslation } from "react-i18next";
+import { LogoCardComponent } from "../../../Shared/LogoCard.component";
 
 export function ContainerPageNotFoundComponent(): JSX.Element {
-    const classes = useStyles();
+    const { t } = useTranslation(["ErrorPage"]);
 
     return (
-        <Card className={classes.root}>
-            <ErrorPageTextContentComponent />
-            <CardMedia
-                component="img"
-                alt="Contemplative Reptile"
-                image="/assets/bfr_logo.gif"
-                title="BfR Logo"
-                css={logoStyle}
-            />
-        </Card>
+        <LogoCardComponent
+            title="404"
+            subtitle={t("Subtitle")}
+            text={t("MainText")}
+        />
     );
 }
